% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh-plot.R
\name{dhsc_style}
\alias{dhsc_style}
\title{dhsc style}
\usage{
dhsc_style()
}
\description{
This function formats a ggplot object. It is mostly copied from the bbc plot package linked below
https://github.com/bbc/bbplot
}
