% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh-plot.R
\name{theme_gov}
\alias{theme_gov}
\title{gov style}
\usage{
theme_gov(base_size = 12, base_colour = "gray40", axes = "x")
}
\description{
This function formats a ggplot object. It is copied from the govstyle package linked below:
https://github.com/ukgovdatascience/govstyle
}
